Add support to export a single database + iam role. Improve config passthrough to the client class to support adding flags easier

+++ dbclient/HiveClient.py
@@ -34,7 +34,7 @@ class HiveClient(dbclient):
                 return x['cluster_id']
         return None
 
-    def launch_cluster(self):
+    def launch_cluster(self, iam_role=None):
         """ Launches a cluster to get DDL statements.
         Returns a cluster_id """
         # removed for now as Spark 3.0 will have backwards incompatible changes
@@ -44,6 +44,11 @@ class HiveClient(dbclient):
         if self.is_aws():
             with open(real_path+'/../data/aws_cluster.json', 'r') as fp:
                 cluster_json = json.loads(fp.read())
+            if iam_role:
+                aws_attr = cluster_json['aws_attributes']
+                print("Creating cluster with: " + iam_role)
+                aws_attr['instance_profile_arn'] = iam_role
+                cluster_json['aws_attributes'] = aws_attr
         else:
             with open(real_path+'/../data/azure_cluster.json', 'r') as fp:
                 cluster_json = json.loads(fp.read())
@@ -104,7 +109,10 @@ class HiveClient(dbclient):
                             json_params=command_payload,
                             version="1.2")
 
-        com_id = command['id']
+        com_id = command.get('id', None)
+        if not com_id:
+            print("ERROR: ")
+            print(command)
         # print('command_id : ' + com_id)
         result_payload = {'clusterId': cid, 'contextId': ec_id, 'commandId': com_id}
 
@@ -171,6 +179,7 @@ class HiveClient(dbclient):
                         else:
                             results['table'] = '{0}.{1}'.format(db_name, table_name)
                             err_log.write(json.dumps(results) + '\n')
+        return True
 
     def check_if_instance_profiles_exists(self, log_file='instance_profiles.log'):
         ip_log = self._export_dir + log_file
@@ -193,6 +202,21 @@ class HiveClient(dbclient):
                     i += 1
             return i
 
+    def export_database(self, db_name, iam_role=None, ms_dir='metastore/'):
+        # check if instance profile exists, ask users to use --users first or enter yes to proceed.
+        start = timer()
+        cid = self.launch_cluster(iam_role)
+        end = timer()
+        print("Cluster creation time: " + str(timedelta(seconds=end - start)))
+        time.sleep(5)
+        ec_id = self.get_execution_context(cid)
+        # if metastore failed log path exists, cleanup before re-running
+        failed_metastore_log_path = self._export_dir + 'failed_metastore.log'
+        if os.path.exists(failed_metastore_log_path):
+            os.remove(failed_metastore_log_path)
+        os.makedirs(self._export_dir + ms_dir + db_name, exist_ok=True)
+        self.log_all_tables(db_name, cid, ec_id, ms_dir, failed_metastore_log_path)
+
     def export_hive_metastore(self, ms_dir='metastore/'):
         # check if instance profile exists, ask users to use --users first or enter yes to proceed.
         instance_profile_log_path = self._export_dir + 'instance_profiles.log'
@@ -212,48 +236,49 @@ class HiveClient(dbclient):
         for db_name in all_dbs:
             self.log_all_tables(db_name, cid, ec_id, ms_dir, failed_metastore_log_path)
 
-        # retry logic below for tables that failed to export during initial export
-        total_failed_entries = self.get_num_of_lines(failed_metastore_log_path)
-        if self.is_aws():
-            if do_instance_profile_exist:
-                print("Instance profiles exist, retrying export of failed tables with each instance profile")
-                err_log_list = []
-                with open(failed_metastore_log_path, 'r') as err_log:
-                    for table in err_log:
-                        err_log_list.append(table)
+        if not self.is_skip_failed():
+            # retry logic below for tables that failed to export during initial export
+            total_failed_entries = self.get_num_of_lines(failed_metastore_log_path)
+            if self.is_aws():
+                if do_instance_profile_exist:
+                    print("Instance profiles exist, retrying export of failed tables with each instance profile")
+                    err_log_list = []
+                    with open(failed_metastore_log_path, 'r') as err_log:
+                        for table in err_log:
+                            err_log_list.append(table)
 
-                with open(instance_profile_log_path, 'r') as iam_log:
-                    for role in iam_log:
-                        role_json = json.loads(role)
-                        iam_role = role_json["instance_profile_arn"]
-                        self.edit_cluster(cid, iam_role)
-                        ec_id = self.get_execution_context(cid)
+                    with open(instance_profile_log_path, 'r') as iam_log:
+                        for role in iam_log:
+                            role_json = json.loads(role)
+                            iam_role = role_json["instance_profile_arn"]
+                            self.edit_cluster(cid, iam_role)
+                            ec_id = self.get_execution_context(cid)
+                            for table in err_log_list:
+                                table_json = json.loads(table)
+                                db_name = table_json['table'].split(".")[0]
+                                table_name = table_json['table'].split(".")[1]
+                                ddl_stmt = 'print(spark.sql("show create table {0}.{1}").collect()[0][0])'.format(db_name,table_name)
+                                results = self.submit_command(cid, ec_id, ddl_stmt)
+                                if results['resultType'] == 'text':
+                                    err_log_list.remove(table)
+                                    with open(self._export_dir + ms_dir + db_name + '/' + table_name, "w") as fp:
+                                        fp.write(results['data'])
+                                else:
+                                    print('failed to get ddl for {0}.{1} with iam role {2}'.format(db_name, table_name, iam_role))
+
+                    os.remove(failed_metastore_log_path)
+                    with open(failed_metastore_log_path, 'w') as fm:
                         for table in err_log_list:
-                            table_json = json.loads(table)
-                            db_name = table_json['table'].split(".")[0]
-                            table_name = table_json['table'].split(".")[1]
-                            ddl_stmt = 'print(spark.sql("show create table {0}.{1}").collect()[0][0])'.format(db_name,table_name)
-                            results = self.submit_command(cid, ec_id, ddl_stmt)
-                            if results['resultType'] == 'text':
-                                err_log_list.remove(table)
-                                with open(self._export_dir + ms_dir + db_name + '/' + table_name, "w") as fp:
-                                    fp.write(results['data'])
-                            else:
-                                print('failed to get ddl for {0}.{1} with iam role {2}'.format(db_name, table_name, iam_role))
-
-                os.remove(failed_metastore_log_path)
-                with open(failed_metastore_log_path, 'w') as fm:
-                    for table in err_log_list:
-                        fm.write(table)
-                failed_count_after_retry = self.get_num_of_lines(failed_metastore_log_path)
-                print("Failed count after retry: " + str(failed_count_after_retry))
+                            fm.write(table)
+                    failed_count_after_retry = self.get_num_of_lines(failed_metastore_log_path)
+                    print("Failed count after retry: " + str(failed_count_after_retry))
+                else:
+                    print("No instance profile to retry export")
+                print("Failed count before retry: " + str(total_failed_entries))
+                print("Total Databases attempted export: " + str(len(all_dbs)))
             else:
-                print("No instance profile to retry export")
-            print("Failed count before retry: " + str(total_failed_entries))
-            print("Total Databases attempted export: " + str(len(all_dbs)))
-        else:
-            print("Failed count: " + str(total_failed_entries))
-            print("Total Databases attempted export: " + str(len(all_dbs)))
+                print("Failed count: " + str(total_failed_entries))
+                print("Total Databases attempted export: " + str(len(all_dbs)))
 
     def create_database_db(self, db_name, ec_id, cid):
         create_db_statement = 'spark.sql("CREATE DATABASE IF NOT EXISTS {0}")'.format(db_name.replace('\n', ''))
